Dataset of BOPs.
This is the dataset we are going to work on. our topic is BOPs.The Balance of Payments Statistics (BOPS) database is like a big collection of information that helps us understand how different countries do business with each other. It tells us things like how much money a country earns from selling things to other countries and how much it spends on buying things from other countries. It also shows us how countries invest in each other. This information is important because it helps governments and businesses make decisions about trade and investments. The BOPS database has data from many countries, and it explains how they collect and organize this information.