Scatter Plot Analysis of Indicators Among Countries.
In the latest update to the repository:

1) In-depth Visualization:
ntroduced scatter plots to analyze the distribution of both Supplementary and Financial indicators across different countries. These visualizations aim to discern patterns, clusters, or any unique characteristics among the countries based on each set of indicators.

Made use of the `ggplot2` package to craft the scatter plots, ensuring clarity and precision in the graphical representation of the data.